feat: Start the curation service with a topic aggregate strategy to handle finding the correct topics
An exaplanation to the implementation from my notes:

We will maintain a count for each topic for each element in a redis db.

We consider a threshold (e.g. `20`) that is the relative weight of a
topic to all topics should be above to be considered.

To implement time decay, we maintain a per topic last time decay
variable. This time decay is used to calculate $\Delta t$. In addition
$$
S_{\text{new}} = S_{\text{old}} * e^{-\lambda * \Delta t}
$$
where $\lambda$ is a parameter to be tuned.

At each input, we add/remove the weight of the user of the topic. Then,
we update the new score for that topic.
After that, we run a calculation to get the scores for the remaining of
topic list to calculate the denominator (without writing, last time
decay variable is updated on write only), and then if there are changes
(any one gets a new threshold), it's to be updated or removed.

We can make sure that the denominator isn't stable yet (< 5.0) to avoid
applying changes